Add check to make sure user is signed in, and owns items or has permission to delete item

+++ src/resolvers/Mutation.js
@@ -46,11 +46,21 @@ const Mutations = {
     );
   },
   async deleteItem(parent, args, ctx, info) {
+    if (!ctx.request.user) throw new Error('You must be signed in!');
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user {id} }`);
     // 2. check if they own it or have permissions
-    // TODO
+    console.log('CTX ', ctx.request.user);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
